+++ doc/changelog.rst
@@ -30,6 +30,10 @@ Added
   gnomonic coordinates. EBSD reference frame documentation.
   (`#204 <https://github.com/pyxem/kikuchipy/pull/204>`_,
   `#215 <https://github.com/pyxem/kikuchipy/pull/215>`_)
+- Geometrical EBSD simulations, projecting a set of Kikuchi bands onto a
+  detector, which can be added to an EBSD signal as markers.
+  (`#204 <https://github.com/pyxem/kikuchipy/pull/204>`_,
+  `#219 <https://github.com/pyxem/kikuchipy/pull/219>`_)
 
 Changed
 -------